Fixed how stats handle time.
Before each stat report was getting system time.  This didn't match up with the simulation speed varying.

Now the environment keeps track of time and the stat reports get the environment time.

+++ src/jevolution/stats/Snapshot.java
@@ -8,9 +8,9 @@ import java.util.Map;
  */
 public class Snapshot {
 	private Map<Class, Double> stats;
-	private long time;
+	private double time;
 
-	Snapshot(long time, Map<Class, Double> stats) {
+	Snapshot(double time, Map<Class, Double> stats) {
 		this.time = time;
 		this.stats = stats;
 	}
@@ -23,7 +23,7 @@ public class Snapshot {
 		return stats.values();
 	}
 
-	public long getTime() {
+	public double getTime() {
 		return time;
 	}
 }
